Fixed support for URIs with query strings
Sound.createInstance was using everything after the last '.' as the file
extension. This meant that any query strings in the URI would be
considered part of the extension and the file would then be considered
to be unsupported.

This has been fixed by using Sound.parsePath to determine whether src
is supported or not. This also resolves the issue where the current
plugin does not support the extension because Sound.parsePath checks
both the supported extensions as well as the current plugin's capabilities.