Adjusted XHR test and added warning reply
The XHR test was giving an error - what do you know, the try catch does not work anymore.  So adjusted the test to just assume it is not XHR if local.  

Also we are getting a warning to not play sounds before interacting.  We are not playing but are testing.  So follow up with a warning that the sounds are ready to play to make whoever is looking at the warnings feel better.